Username availability
Sorry for using a pull request to contact you. I see you are not using this GitHub name and we are interested in getting it because it corresponds to our company name. We have several open-source projects we would like to put on sqlbi main repository, now we use sql-bi but several people are confused by that.
Please, let me know if you want to consider the name change and evaluate a compensation for that.
I look forward to hearing from you.
Thanks,

Marco Russo